refactor: move FlixError into its own class generator
The generated `FlixError` class lived inside `ClassConstants`, next to the
well-known members of JDK classes, and was the only entry there that
generates bytecode. It now has its own generator in `classes/GenFlixError`,
matching the four error classes that extend it.
